used only one requestData function for all data type
